Split out tests and fix timed/default cases
